chore(compact-js): Refactor command and CLI option usage (#157)
Improves the internal error handling around (de)serialization of input files (i.e., the state and Zswap local state files). Also renames the command line options representing input files so that they are the named equivalent of the outputs. For example, we now use `'--input-ps'` when supplying the input private state data, and use `'--output-ps'` to obtain the equivalent output.